#1 First version - adding a persistence test for Watchdog entity; building the database via DatabaseBuilder

+++ src/Core/Features/Shared/Domain/VersionedEntity.cs
@@ -13,5 +13,5 @@ namespace MrWatchdog.Core.Features.Shared.Domain;
 /// </summary>
 public abstract class VersionedEntity : Entity<long>
 {
-    public virtual int Version { get; protected set; }
+    public virtual long Version { get; protected set; }
 }
